frontend: switch to parcel v2
As always frontend dev is convoluted. parcel v1 doesn't play nicely
with vue: https://github.com/parcel-bundler/parcel/issues/2408 so
we would need to disable source maps. But that'll slow down dev, so
we migrate to v2. That comes with its own baggage, but hopefully
it'll be the lesser of the two evils.